refactored fixtures
