+++ tests/tests_games/test_local_xai.py
@@ -197,9 +197,9 @@ def test_sentiment_classifier(mask_strategy):
         )
 
 
-def test_resnet_model_class(test_image_and_path):
+def test_resnet_model_class(image_and_path):
     """Tests the creation of the ResNet Model class."""
-    test_image, _ = test_image_and_path
+    test_image, _ = image_and_path
     resnet_model = ResNetModel(n_superpixels=14, input_image=test_image, verbose=True, batch_size=2)
     assert resnet_model.n_superpixels == 14
 
@@ -218,9 +218,9 @@ def test_resnet_model_class(test_image_and_path):
     assert output[1] != 0.0
 
 
-def test_image_classifier_game_resnet(test_image_and_path):
+def test_image_classifier_game_resnet(image_and_path):
     """Tests the ImageClassifierGame with the ResNet models."""
-    test_image, path_from_test_root = test_image_and_path
+    test_image, path_from_test_root = image_and_path
     game = ImageClassifierLocalXAI(
         model_name="resnet_18", verbose=True, x_explain_path=path_from_test_root
     )
@@ -240,9 +240,9 @@ def test_image_classifier_game_resnet(test_image_and_path):
     assert np.isclose(game_small(game_small.grand_coalition), grand_coal_output)
 
 
-def test_vit_model_class(test_image_and_path):
+def test_vit_model_class(image_and_path):
     """Tests the creation of the ViTModel class."""
-    test_image, _ = test_image_and_path
+    test_image, _ = image_and_path
     vit_model = ViTModel(n_patches=16, input_image=test_image, verbose=False)
     assert vit_model.n_patches == 16
     assert vit_model(coalitions=np.ones(16))[0] >= 0.9195  # check that call works
@@ -255,9 +255,9 @@ def test_vit_model_class(test_image_and_path):
         ViTModel(n_patches=10, input_image=test_image)
 
 
-def test_image_classifier_game_vit(test_image_and_path):
+def test_image_classifier_game_vit(image_and_path):
     """Tests the ImageClassifierGame with the ViT models."""
-    test_image, path_from_test_root = test_image_and_path
+    test_image, path_from_test_root = image_and_path
     game = ImageClassifierLocalXAI(
         x_explain_path=path_from_test_root, model_name="vit_9_patches", normalize=True, verbose=True
     )
